backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-cloud-vision==3.5.0
google-generativeai==0.3.2
PyPDF2==3.0.1
aiohttp==3.9.3
python-multipart==0.0.9
pydantic==1.10.13
aiofiles==23.2.1
